remove op prompt for view records

+++ CodingTracker.Kaylubr/Views/UserInterface.cs
@@ -20,11 +20,6 @@ internal static class UserInterface
              .AddChoices(Enum.GetValues<MenuChoices>())
             );
 
-            if (!Helper.Confirmation("Do you want to do the operation?"))
-            {
-                continue;
-            }
-
             switch (choice)
             {
                 case MenuChoices.View:
@@ -32,11 +27,23 @@ internal static class UserInterface
                     break;
                 case MenuChoices.Insert:
                     CodingTrackerController.InsertSession();
+
+                    if (!Helper.Confirmation("Proceed with this operation? (No will return to the main menu)"))
+                        continue;
+
                     break;
                 case MenuChoices.Update:
                     CodingTrackerController.UpdateRecord();
+
+                    if (!Helper.Confirmation("Proceed with this operation? (No will return to the main menu)"))
+                        continue;
+
                     break;
                 case MenuChoices.Delete:
+
+                    if (!Helper.Confirmation("Proceed with this operation? (No will return to the main menu)"))
+                        continue;
+
                     CodingTrackerController.DeleteRecord();
                     break;
             }
